+++ tests/_support/Helper/Integration.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Helper;
+
+use Codeception\Module;
+
+/**
+ * Custom actions for the integration suite. All public methods declared here
+ * are available on `$this->tester`.
+ */
+class Integration extends Module
+{
+}